# perf(Math): Avoid Math.min, Math.max in vtkMath.arrayMin, vtkMath.arr…