--- a/Sources/Common/Core/Math/index.js
+++ b/Sources/Common/Core/Math/index.js
@@ -56,7 +56,7 @@ function arrayMin(arr) {
 function arrayMax(arr) {
   let maxValue = -Infinity;
   for (let i = 0, len = arr.length; i < len; ++i) {
-    if (arr[i] > maxValue) {
+    if (maxValue < arr[i]) {
       maxValue = arr[i];
     }
   }