…ayMax

Timings:

  Existing Math.min: ~210ms for arrayMin, ~2712ms for entire volume rendering initialization
  Comparison operator: ~35ms for arrayMin, ~2070ms for entire volume rendering initialization
  Ternary operator: ~37ms for arrayMin, ~2125ms for entire volume rendering initialization

CC: @martinken @jourdain @agirault