@agirault 2018-02-12
Will the `perf` trigger a new release? Because I want a new release :heart: