@jourdain 2018-02-13
No but since I merged it with other fixes, you will get you new release...