@jourdain 2018-02-13
Actually looking at the ChangeLog, perf() does create a new release... ;-)